track id: #113: Code changes related to RFC 5 and 8: This code drop adds the implementation of the functions NullValue, ToDouble, ToFloat, ToInt32 and ToInt64 and modifies some other function implementation with respect to exception handling (in this one exception message is replaced with a different one).
git-svn-id: http://svn.osgeo.org/fdo/trunk@3327 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Utilities/ExpressionEngine/Src/Functions/String/FdoFunctionTrim.cpp
@@ -160,7 +160,7 @@ FdoLiteralValue *FdoFunctionTrim::Evaluate (
         if (pos == string_length)
             return FdoStringValue::Create();
         else
-          base_string = base_string.Mid((size_t) pos, string_length);
+          base_string = base_string.Mid((size_t) pos, (size_t) string_length);
 
     }  //  if ((FdoCommonStringUtil::StringCompareNoCase( ...
